requirements.txt
asn1crypto==0.24.0
boto3==1.9.162
botocore==1.12.163
certifi==2019.3.9
cffi==1.12.3
Click==7.0
cryptography==2.7
docutils==0.14
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.3
Werkzeug==0.15.4
win-inet-pton==1.1.0
wincertstore==0.2
